alligator/pyproject.toml
[tool.poetry]
name = "alligator"
version = "0.1.0"
description = ""
authors = ["Katun Ilya <katun.ilya@gmail.com>"]
readme = "README.md"
packages = [{include = "alligator"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
pydantic = {extras = ["email"], version = "^1.10.2"}
uvicorn = "^0.19.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
toolz = "^0.12.0"
python-dotenv = "^0.21.0"
pymongo = "^4.3.2"
scipy = "^1.9.3"
numpy = "^1.23.5"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
pytest = "^7.1.3"
isort = "^5.10.1"
pytest-asyncio = "^0.20.1"
flake8-bugbear = "^22.9.23"
flake8-comprehensions = "^3.10.0"
flake8-docstrings = "^1.6.0"
flake8-eradicate = "^1.4.0"
flake8-isort = "^5.0.0"
pep8-naming = "^0.13.2"
flake8-string-format = "^0.3.0"
flake8-bandit = "^4.1.1"
flake8-builtins = "^2.0.0"
flake8-black = "^0.3.3"
flake8-requirements = "^1.7.2"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
